nvprof cold start measurements work with testModel. Also moves getNGpu and friends to infbench rather than benchmark/util.py
